# ebook-converter/src/app/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BookConv API",
    "description": "Ebook format conversion API powered by Calibre. Supports 28+ formats including EPUB, MOBI, AZW3, PDF, DOCX, TXT, FB2, LIT, RTF, and more.\n\n**Base URL:** `https://bookconv.com/api`",
    "version": "1.0.0",
    "contact": {
      "name": "BookConv Support",
      "url": "https://bookconv.com"
    }
  },
  "servers": [
    {
      "url": "https://bookconv.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer token for authenticated users (optional). Free users are rate-limited to 20 requests/minute per IP. Pro users get priority processing."
      }
    },
    "schemas": {
...
      "BatchFileItem": {
        "type": "object",
        "properties": {
          "index": { "type": "integer" },
          "name": { "type": "string" },
          "size": { "type": "integer" },
          "sourceFormat": { "type": "string" },
          "targetFormat": { "type": "string" },
          "jobId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["queued", "processing", "completed", "failed"] },
          "error": { "type": "string", "nullable": true },
          "result": { "$ref": "#/components/schemas/BatchFileResult", "nullable": true }
        }
      },
      "BatchFileResult": {
        "type": "object",
        "properties": {
          "extension": { "type": "string" },
          "mimeType": { "type": "string" },
          "fileSize": { "type": "integer" }
        }
      },
      "BatchSummary": {
        "type": "object",
        "properties": {
          "batchId": { "type": "string", "format": "uuid" },
          "total": { "type": "integer" },
          "queued": { "type": "integer" },
          "processing": { "type": "integer" },
          "completed": { "type": "integer" },
          "failed": { "type": "integer" },
          "files": { "type": "array", "items": { "$ref": "#/components/schemas/BatchFileItem" } },
          "zipUrl": { "type": "string", "nullable": true },
          "message": { "type": "string" }
        }
      },
      "BatchStatusResponse": {
        "type": "object",
        "properties": {
          "batchId": { "type": "string" },
          "targetFormat": { "type": "string" },
          "total": { "type": "integer" },
          "completed": { "type": "integer" },
          "failed": { "type": "integer" },
          "processing": { "type": "integer" },
          "queued": { "type": "integer" },
          "overallProgress": { "type": "integer", "minimum": 0, "maximum": 100 },
          "files": { "type": "array", "items": { "$ref": "#/components/schemas/BatchFileItem" } },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string", "nullable": true }
        }
      }
    },
    "parameters": {
...
      "BatchIdPath": {
        "name": "batchId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "The batch ID returned from the batch convert endpoint"
      }
    },
    "responses": {
      "RateLimited": {
        "description": "Rate limit exceeded (429)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Too many requests. Please try again later." }
          }
        },
        "headers": {
          "X-RateLimit-Limit": { "schema": { "type": "integer" }, "description": "Max requests per window" },
          "X-RateLimit-Remaining": { "schema": { "type": "integer" }, "description": "Remaining requests" },
          "X-RateLimit-Reset": { "schema": { "type": "integer" }, "description": "Unix timestamp when the window resets" },
          "Retry-After": { "schema": { "type": "integer" }, "description": "Seconds to wait before retrying" }
        }
      },
...
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
...
    "/convert/batch": {
      "post": {
        "tags": ["Batch Conversion"],
        "summary": "Submit multiple files for batch conversion",
        "description": "Upload up to 50 files for parallel conversion. Each file is processed independently. Returns a batchId and per-file status.",
        "operationId": "batchConvert",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["files"],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" },
                    "minItems": 1,
                    "maxItems": 50,
                    "description": "Multiple ebook files to convert (max 50 files, each up to 50 MB)"
                  },
                  "target_format": {
                    "type": "string",
                    "description": "Target format (defaults to epub)",
                    "example": "epub"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch submitted — some or all files converted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchSummary" },
                "example": {
                  "batchId": "b1c2d3e4-f5a6-7890-bcde-f12345678901",
                  "total": 3,
                  "queued": 0,
                  "processing": 0,
                  "completed": 2,
                  "failed": 1,
                  "files": [
                    { "index": 0, "name": "book1.epub", "size": 524288, "sourceFormat": "epub", "status": "completed", "error": null, "result": { "extension": "azw3", "mimeType": "application/x-mobipocket-ebook", "fileSize": 480000 } },
                    { "index": 1, "name": "book2.pdf", "size": 1048576, "sourceFormat": "pdf", "status": "completed", "error": null, "result": { "extension": "azw3", "mimeType": "application/x-mobipocket-ebook", "fileSize": 960000 } },
                    { "index": 2, "name": "bad.xyz", "size": 1024, "sourceFormat": "xyz", "status": "failed", "error": "Unsupported format: .xyz", "result": null }
                  ],
                  "zipUrl": "/api/convert/batch/b1c2d3e4-f5a6-7890-bcde-f12345678901/download",
                  "message": "2/3 files converted successfully"
                }
              }
            }
          },
          "207": {
            "description": "Multi-status — batch submitted, results being processed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchSummary" },
                "example": {
                  "batchId": "b1c2d3e4-f5a6-7890-bcde-f12345678901",
                  "total": 5,
                  "queued": 5,
                  "processing": 0,
                  "completed": 0,
                  "failed": 0,
                  "files": [],
                  "zipUrl": null,
                  "message": "Batch submitted for processing"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "No files provided" }
              }
            }
          },
          "413": {
            "description": "Too many files or file too large",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Maximum 50 files per batch (Pro limit)" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      },
      "get": {
        "tags": ["Batch Conversion"],
        "summary": "Get batch conversion status",
        "description": "Query the current status of a batch conversion by batchId.",
        "operationId": "getBatchStatus",
        "parameters": [
          {
            "name": "batchId",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The batch ID from the POST response"
          }
        ],
        "responses": {
          "200": {
            "description": "Batch status retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchStatusResponse" },
                "example": {
                  "batchId": "b1c2d3e4-f5a6-7890-bcde-f12345678901",
                  "targetFormat": "epub",
                  "total": 5,
                  "completed": 3,
                  "failed": 1,
                  "processing": 1,
                  "queued": 0,
                  "overallProgress": 80,
                  "files": [],
                  "createdAt": "2026-07-17T10:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Missing batchId parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Missing batchId query parameter" }
              }
            }
          },
          "404": {
            "description": "Batch not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Batch not found" }
              }
            }
          }
        }
      }
    },
    "/convert/batch/{batchId}/download": {
      "get": {
        "tags": ["Batch Conversion"],
        "summary": "Download batch results as ZIP",
        "description": "Download all successfully converted files from a batch as a single ZIP archive.",
        "operationId": "downloadBatchZip",
        "parameters": [{ "$ref": "#/components/parameters/BatchIdPath" }],
        "responses": {
          "200": {
            "description": "ZIP archive of converted files",
            "content": {
              "application/zip": {
                "schema": { "type": "string", "format": "binary" }
              }
            },
            "headers": {
              "Content-Disposition": { "schema": { "type": "string" }, "example": "attachment; filename=\"converted-a1b2c3d4.zip\"" },
              "X-Batch-Id": { "schema": { "type": "string" } },
              "X-Completed-Files": { "schema": { "type": "integer" } }
            }
          },
          "404": {
            "description": "Batch not found or no completed files",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "examples": {
                  "notFound": { "value": { "error": "Batch not found" } },
                  "noCompleted": { "value": { "error": "No completed files to download" } }
                }
              }
            }
          }
        }
      }
    },
...
  }
}